Tracking Cam
Camera now lerps to the active player's position, it uses local variables in its functions just in case it needs to be moved somewhere else. Made most sense to be in G manager instead of map because its Having a reference to camera through finding its tag vs Passing active players to map in code, I went with the former as it seemed simpler.