quick late night commit
made it so left shift toggles between the cameras default position (map view) and the tracking cam. Changed map start to awake so it sets the cameras transform for it to be properly set when game manager pulls it.

will look to add a CameraManager and I'll see if I can fine tune the cameras map view/default pos to the new map hopefully without hard code tomorrow.