Fixed combat ending too early and now the flow is back with monsters included.
Just gotta work out how to differentiate between hit and dead, because anims atm are triggered before damage calc so its a bit tricky. Might make it so hit is the default anim triggered, then death OR run away is triggered based on damage clac. But id rather not do it that way and do Dead or Hit > followed by run instead.