Goblin now does run away anims, still need to test dead, but that should work
Hit is now the default reaction to a loss and the death or run away anim is followed after, may change.

messed around with some animation transitions by unticking their exit times so they transitioned immediately to where they need to be more. Fixed some anims not having events too.

should be good to add in the rest of the monsters next

Signed-off-by: JackMays <maysey43@hotmail.com>